+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0 - 2023-01-30
 
-- Added `bundleDts()` function with `tsconfigJsonPath`, `entry`, `include`, `exclude` and `outFile` options
+- Added `bundleDts()` function with `tsconfigFilePath`, `entry`, `include`, `exclude` and `outFile` options